+++ app/layout.tsx
@@ -1,8 +1,9 @@
 import "./globals.css";
 
+import { Metadata } from "next";
 import { Inter, Space_Grotesk } from "next/font/google";
 
-import siteMetadata from "@/lib/metadata";
+import siteMetadata, { BASE_URL, defaultAuthor } from "@/lib/metadata";
 import { Toaster } from "@/components/ui/toaster";
 import { Analytics } from "@/components/analytics";
 import { BackTopButton } from "@/components/back-to-top";
@@ -16,9 +17,17 @@ const spaceGrotesk = Space_Grotesk({
 });
 const inter = Inter({ subsets: ["latin"], display: "swap", variable: "--font-inter" });
 
-export const metadata = {
+export const metadata: Metadata = {
+  metadataBase: new URL(BASE_URL),
   title: siteMetadata.title,
   description: siteMetadata.description,
+  authors: [{ name: defaultAuthor.name, url: defaultAuthor.website }],
+  alternates: {
+    canonical: "./",
+    types: {
+      "application/rss+xml": `${BASE_URL}/feed.xml`,
+    },
+  },
 };
 
 interface RootLayoutProps {
@@ -29,7 +38,7 @@ export default function RootLayout({ children }: RootLayoutProps) {
   return (
     <html lang="en" className={`${spaceGrotesk.variable} ${inter.variable}`} suppressHydrationWarning>
       <body className="min-h-screen bg-gradient-to-b from-slate-100 to-white text-slate-900 antialiased dark:bg-gradient-to-b dark:from-gray-900 dark:to-gray-800 dark:text-slate-50">
-        <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
+        <ThemeProvider attribute="class" defaultTheme={siteMetadata.defaultTheme} enableSystem>
           {children}
           <BackTopButton />
           <Toaster />