+++ lib/content-definitions/author.ts
@@ -1,0 +1,9 @@
+import { defineNestedType } from "contentlayer/source-files";
+
+export const Author = defineNestedType(() => ({
+  name: "Author",
+  fields: {
+    name: { type: "string", required: true },
+    image: { type: "string" },
+  },
+}));